requirements.txt
fitz
python-docx